Resolve #3258: propagate Node response compression failures
