+++ .changeset/fix-runtime-node-compression-failures.md
@@ -1,0 +1,5 @@
+---
+"@fluojs/runtime": patch
+---
+
+Propagate failed Node response compression writes and terminate responses whose compression stream closes before completion.
