+++ packages/runtime/src/internal-node.ts
@@ -1,2 +1,2 @@
 export * from './node/internal-node.js';
-export { createNodeEarlyHintsCapability } from './node/internal-node-response.js';
+export { createNodeEarlyHintsCapability } from './node/internal-node-early-hints.js';
